infrastructure/lambda/claude_proxy/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
psycopg2-binary==2.9.9
